requirements.txt
pandas
SQLAlchemy
python-dotenv
aiohttp
matplotlib
PyMySQL
cryptography